moved all images from images folder under componets to public
