+++ Dental-House-UI/public/images/cavity.png
