+++ Dental-House-UI/src/Components/HeaderAndFooter/Header.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import logo from "../images/logo.jpg";
+import logo from "../../../public/images/logo.jpg";
 
 const Header = ({ scrollToSection, refs }) => {
   return (
